# packages/schemas/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Eyohe OSINT API",
  "description": "Public Intelligence. Connected Evidence. \u2014 local-first OSINT investigation platform.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/cases/{case_id}/evidence": {
   "get": {
    "tags": [
     "evidence"
    ],
    "summary": "List Evidence",
    "operationId": "list_evidence_api_v1_cases__case_id__evidence_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "evidence_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Evidence Type"
      }
     },
     {
      "name": "review_state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Review State"
      }
     },
     {
      "name": "confidence",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Confidence"
      }
     },
     {
      "name": "source_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Source Id"
      }
     },
     {
      "name": "collector",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Collector"
      }
     },
     {
      "name": "entity_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Id"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_EvidenceOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "evidence"
    ],
    "summary": "Create Evidence",
    "description": "Analyst-created evidence (e.g. from the query playground or manual review).",
    "operationId": "create_evidence_api_v1_cases__case_id__evidence_post",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EvidenceCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EvidenceDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/evidence/{evidence_id}": {
   "get": {
    "tags": [
     "evidence"
    ],
    "summary": "Get Evidence",
    "operationId": "get_evidence_api_v1_evidence__evidence_id__get",
    "parameters": [
     {
      "name": "evidence_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Evidence Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EvidenceDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/evidence/{evidence_id}/review": {
   "post": {
    "tags": [
     "evidence"
    ],
    "summary": "Review Evidence",
    "operationId": "review_evidence_api_v1_evidence__evidence_id__review_post",
    "parameters": [
     {
      "name": "evidence_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Evidence Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReviewRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EvidenceDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/evidence/{evidence_id}/artifacts/{artifact_id}": {
   "get": {
    "tags": [
     "evidence"
    ],
    "summary": "Download Artifact",
    "operationId": "download_artifact_api_v1_evidence__evidence_id__artifacts__artifact_id__get",
    "parameters": [
     {
      "name": "evidence_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Evidence Id"
      }
     },
     {
      "name": "artifact_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Artifact Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/cases/{case_id}/sources": {
   "get": {
    "tags": [
     "evidence"
    ],
    "summary": "List Sources",
    "operationId": "list_sources_api_v1_cases__case_id__sources_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "source_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Source Type"
      }
     },
     {
      "name": "collector",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Collector"
      }
     },
     {
      "name": "max_tier",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer",
         "maximum": 5,
         "minimum": 1
        },
        {
         "type": "null"
        }
       ],
       "title": "Max Tier"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_SourceOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sources": {
   "get": {
    "tags": [
     "evidence"
    ],
    "summary": "List All Sources",
    "operationId": "list_all_sources_api_v1_sources_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_SourceOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sources/{source_id}": {
   "get": {
    "tags": [
     "evidence"
    ],
    "summary": "Get Source",
    "operationId": "get_source_api_v1_sources__source_id__get",
    "parameters": [
     {
      "name": "source_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Source Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SourceDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sources/{source_id}/snapshots/{snapshot_id}": {
   "get": {
    "tags": [
     "evidence"
    ],
    "summary": "Get Snapshot",
    "operationId": "get_snapshot_api_v1_sources__source_id__snapshots__snapshot_id__get",
    "parameters": [
     {
      "name": "source_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Source Id"
      }
     },
     {
      "name": "snapshot_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Snapshot Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SnapshotDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sources/{source_id}/compare": {
   "get": {
    "tags": [
     "evidence"
    ],
    "summary": "Compare Snapshots",
    "description": "Diff the two most recent snapshots of a source (title/text changes).",
    "operationId": "compare_snapshots_api_v1_sources__source_id__compare_get",
    "parameters": [
     {
      "name": "source_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Source Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Compare Snapshots Api V1 Sources  Source Id  Compare Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "EvidenceCreate": {
    "properties": {
     "claim": {
      "type": "string",
      "maxLength": 4000,
      "minLength": 1,
      "title": "Claim"
     },
     "evidence_type": {
      "type": "string",
      "title": "Evidence Type",
      "default": "DIRECT_STATEMENT"
     },
     "source_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Id"
     },
     "source_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Url"
     },
     "excerpt": {
      "type": "string",
      "title": "Excerpt",
      "default": ""
     },
     "context": {
      "type": "string",
      "title": "Context",
      "default": ""
     },
     "observed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Observed At"
     },
     "entity_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Entity Ids"
     }
    },
    "type": "object",
    "required": [
     "claim"
    ],
    "title": "EvidenceCreate"
   },
   "EvidenceDetail": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "display_id": {
      "type": "string",
      "title": "Display Id"
     },
     "case_id": {
      "type": "string",
      "format": "uuid",
      "title": "Case Id"
     },
     "investigation_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Investigation Id"
     },
     "source_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Id"
     },
     "snapshot_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Snapshot Id"
     },
     "evidence_type": {
      "type": "string",
      "title": "Evidence Type"
     },
     "claim": {
      "type": "string",
      "title": "Claim"
     },
     "excerpt": {
      "type": "string",
      "title": "Excerpt"
     },
     "excerpt_verified": {
      "type": "boolean",
      "title": "Excerpt Verified"
     },
     "context": {
      "type": "string",
      "title": "Context"
     },
     "collector": {
      "type": "string",
      "title": "Collector"
     },
     "collection_method": {
      "type": "string",
      "title": "Collection Method"
     },
     "collected_at": {
      "type": "string",
      "format": "date-time",
      "title": "Collected At"
     },
     "observed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Observed At"
     },
     "confidence": {
      "type": "string",
      "title": "Confidence"
     },
     "review_state": {
      "type": "string",
      "title": "Review State"
     },
     "reviewed_by": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed By"
     },
     "reviewed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed At"
     },
     "review_note": {
      "type": "string",
      "title": "Review Note"
     },
     "content_hash": {
      "type": "string",
      "title": "Content Hash"
     },
     "structured": {
      "additionalProperties": true,
      "type": "object",
      "title": "Structured"
     },
     "entity_ids": {
      "items": {},
      "type": "array",
      "title": "Entity Ids"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo"
     },
     "redacted": {
      "type": "boolean",
      "title": "Redacted"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "source": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/SourceOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "artifacts": {
      "items": {
       "$ref": "#/components/schemas/ArtifactOut"
      },
      "type": "array",
      "title": "Artifacts"
     },
     "entities": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Entities"
     },
     "findings": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Findings"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_id",
     "case_id",
     "investigation_id",
     "source_id",
     "snapshot_id",
     "evidence_type",
     "claim",
     "excerpt",
     "excerpt_verified",
     "context",
     "collector",
     "collection_method",
     "collected_at",
     "observed_at",
     "confidence",
     "review_state",
     "reviewed_by",
     "reviewed_at",
     "review_note",
     "content_hash",
     "structured",
     "entity_ids",
     "is_demo",
     "redacted",
     "created_at"
    ],
    "title": "EvidenceDetail"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Page_EvidenceOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/EvidenceOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page",
      "default": 1
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size",
      "default": 50
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[EvidenceOut]"
   },
...
   "Page_SourceOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/SourceOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page",
      "default": 1
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size",
      "default": 50
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "Page[SourceOut]"
   },
...
   "ReviewRequest": {
    "properties": {
     "state": {
      "type": "string",
      "pattern": "^(ACCEPTED|REJECTED|NEEDS_VERIFICATION|PENDING)$",
      "title": "State"
     },
     "note": {
      "type": "string",
      "maxLength": 4000,
      "title": "Note",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "state"
    ],
    "title": "ReviewRequest"
   },
...
   "SourceDetail": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "display_id": {
      "type": "string",
      "title": "Display Id"
     },
     "case_id": {
      "type": "string",
      "format": "uuid",
      "title": "Case Id"
     },
     "source_type": {
      "type": "string",
      "title": "Source Type"
     },
     "url": {
      "type": "string",
      "title": "Url"
     },
     "canonical_url": {
      "type": "string",
      "title": "Canonical Url"
     },
     "domain": {
      "type": "string",
      "title": "Domain"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "publisher": {
      "type": "string",
      "title": "Publisher"
     },
     "author": {
      "type": "string",
      "title": "Author"
     },
     "published_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Published At"
     },
     "collected_at": {
      "type": "string",
      "format": "date-time",
      "title": "Collected At"
     },
     "collector": {
      "type": "string",
      "title": "Collector"
     },
     "tier": {
      "type": "integer",
      "title": "Tier"
     },
     "reliability_note": {
      "type": "string",
      "title": "Reliability Note"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo"
     },
     "metadata": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metadata"
     },
     "snapshots": {
      "items": {
       "$ref": "#/components/schemas/SnapshotOut"
      },
      "type": "array",
      "title": "Snapshots"
     },
     "evidence_count": {
      "type": "integer",
      "title": "Evidence Count",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_id",
     "case_id",
     "source_type",
     "url",
     "canonical_url",
     "domain",
     "title",
     "publisher",
     "author",
     "published_at",
     "collected_at",
     "collector",
     "tier",
     "reliability_note",
     "status",
     "is_demo"
    ],
    "title": "SourceDetail"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}